# internal/api/handler/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.createEphemeralTokenRequest:
    properties:
      name:
        type: string
    type: object
...
  handler.createWorkerRequest:
    properties:
      name:
        type: string
    type: object
...
  handler.updateWorkerRequest:
    properties:
      hwaccel:
        type: string
      threads:
        type: integer
    type: object
...
  models.EphemeralWorkerToken:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
  models.TranscodeWorker:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      hwaccel:
        type: string
      id:
        type: string
      is_ephemeral:
        type: boolean
      last_heartbeat:
        type: string
      name:
        type: string
      status:
        type: string
      threads:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST & WebSocket API specification for Prism, a self-hosted media server.
  title: Prism Media Server API
  version: "1.0"
paths:
...
  /worker-tokens:
    get:
      description: Retrieve a list of all active ephemeral worker registration tokens.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EphemeralWorkerToken'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Ephemeral Worker Tokens
      tags:
      - Worker Administration
    post:
      consumes:
      - application/json
      description: Create a new re-usable registration token for ephemeral workers.
      parameters:
      - description: Token parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createEphemeralTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EphemeralWorkerToken'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Ephemeral Worker Token
      tags:
      - Worker Administration
  /worker-tokens/{id}:
    delete:
      description: Revoke/delete an ephemeral worker registration token.
      parameters:
      - description: Token ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Token not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Ephemeral Worker Token
      tags:
      - Worker Administration
...
  /workers:
    get:
      description: Retrieve a list of all registered remote transcode workers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TranscodeWorker'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Transcode Workers
      tags:
      - Worker Administration
    post:
      consumes:
      - application/json
      description: Register a new remote transcode worker. Returns the created worker
        metadata including its authentication secret key.
      parameters:
      - description: Worker parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createWorkerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TranscodeWorker'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register Transcode Worker
      tags:
      - Worker Administration
  /workers/{id}:
    delete:
      description: De-register a remote transcode worker and invalidate its API credentials.
      parameters:
      - description: Worker ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Worker not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Transcode Worker
      tags:
      - Worker Administration
    put:
      consumes:
      - application/json
      description: Update concurrency limits (threads) or hardware acceleration configuration
        for a registered worker.
      parameters:
      - description: Worker ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Worker settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateWorkerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TranscodeWorker'
        "400":
          description: Invalid ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Worker not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Transcode Worker Settings
      tags:
      - Worker Administration
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
  WorkerAuth:
    description: The secret API Key assigned to the transcode worker.
    in: header
    name: X-Worker-API-Key
    type: apiKey
swagger: "2.0"